fix: don't emit examples larger than 500 lines
Currently, the `aws-quicksight` library would produce an example that is
30,000 lines. There is no real point to this and in fact blows up the
size of the Python source file by quite a lot, so refuse to generate
examples that are too large.

In the future we may choose to do this eliding more intelligently (for
example, by skipping optional properties and/or limiting recursion
depth), but this is a multidimensional optimization problem that may
not be so simple to solve as it seems at first. Most important thing
is to stop the bleeding.