fix: generates wrong imports if filename doesn't match submodule name
This tool generates imports for submodules in the following way:

```ts
// Single level submodules
import { level1 } from 'library';
var x: level1.Type;

// Nested submodules
import { level2 } from 'library/level1';
var x: level2.Type;

// Which is better than the alternative
import { level1 } from 'library';
var x: level1.level2.Type;
```

But notice here:

```ts
import { level2 } from 'library/level1';
//                             ^^^^^^^^ need to guess at this
```

The filename of the first-level submodule that exports the second-level
submodule is not known by jsii, since it assumes that all symbols are
always accessed via expression chaining from the export root.

This file name is fully outside the domain of jsii, and controlled by
the interplay of JavaScript filenames and what's in the `"exports"`
section of `package.json`.

`cdk-generate-synthetic-examples` would therefore guess, and guess that
the filename was the same as the identifier name. Usually true, but not
in the case of AWS service names: `aws-s3` vs `aws_s3`.

The change
==========

In this PR, use the `symbolId` of a submodule (which they luckily have!)
to find the actual on-disk filename that defines the submodule entry
point, and correlate that with the `"exports"` table of `package.json`
to come up with the import statement to generate.

Hypothetical example:

```ts
// Submodule:     library.level_1
//                          ^^^-------- incorrect filename guess

const symbolId = "lib/level-1/index:";
//                  ^^^^^-------------- take this

// package.json
{
  "exports": {
    "./level-1": "./lib/level-1/index.js"
                      ^^^^^^------------- match against this
  }
}

// Final import:

import { level2 } from 'library/level-1';
```

In order to be able to read the `package.json` of an assembly,
we need to know where it lives on disk, so we introduce a
`DirTrackingTypeSystem` as a subclass of `reflect.TypeSystem` to
keep track of the directories we are loading assemblies from.
(This feature will come out properly in `jsii-reflect` soon, but
we've got it earlier here).